+++ src/lib/services/GroupsService.ts
@@ -776,8 +776,8 @@ const LOG_ID = "GROUPS/SVCE - ";
 
             that._eventEmitter.emit("evt_internal_groupcreated", groupCreated);
         }).catch((err) => {
-            that._logger.log("error", LOG_ID + "(_onGroupCreated) Error.");
-            that._logger.log("internalerror", LOG_ID + "(_onGroupCreated) Error : ", err);
+            that._logger.log("warn", LOG_ID + "(_onGroupCreated) Error : ", err);
+            //that._logger.log("internalerror", LOG_ID + "(_onGroupCreated) Error : ", err);
         });
     }
 
@@ -828,8 +828,8 @@ const LOG_ID = "GROUPS/SVCE - ";
 
             that._eventEmitter.emit("evt_internal_groupupdated", groupUpdated);
         }).catch((err) => {
-            that._logger.log("error", LOG_ID + "(_onGroupUpdated) Error.");
-            that._logger.log("internalerror", LOG_ID + "(_onGroupUpdated) Error : ", err);
+            that._logger.log("warn", LOG_ID + "(_onGroupUpdated) Error : ", err);
+            //that._logger.log("internalerror", LOG_ID + "(_onGroupUpdated) Error : ", err);
         });
     }
 
@@ -859,8 +859,8 @@ const LOG_ID = "GROUPS/SVCE - ";
 
             that._eventEmitter.emit("evt_internal_useraddedingroup", groupUpdated, contact);
         }).catch((err) => {
-            that._logger.log("error", LOG_ID + "(_onUserAddedInGroup) Error.");
-            that._logger.log("internalerror", LOG_ID + "(_onUserAddedInGroup) Error : ", err);
+            that._logger.log("warn", LOG_ID + "(_onUserAddedInGroup) Error : ", err);
+            //that._logger.log("internalerror", LOG_ID + "(_onUserAddedInGroup) Error : ", err);
         });
     }
 
@@ -892,8 +892,8 @@ const LOG_ID = "GROUPS/SVCE - ";
 
             that._eventEmitter.emit("evt_internal_userremovedfromgroup", groupUpdated, contact);
         }).catch((err) => {
-            that._logger.log("error", LOG_ID + "(_onUserRemovedFromGroup) Error.");
-            that._logger.log("internalerror", LOG_ID + "(_onUserRemovedFromGroup) Error : ", err);
+            that._logger.log("warn", LOG_ID + "(_onUserRemovedFromGroup) Error : ", err);
+            //that._logger.log("internalerror", LOG_ID + "(_onUserRemovedFromGroup) Error : ", err);
         });
     }